bug fixed

+++ src/main/java/com/nirmal/personalfinancetracker/service/impl/ExpenseServiceImpl.java
@@ -1,11 +1,15 @@
 package com.nirmal.personalfinancetracker.service.impl;
 
 import com.nirmal.personalfinancetracker.dto.request.AddExpenseDto;
+import com.nirmal.personalfinancetracker.dto.response.ExpenseResponseDto;
+import com.nirmal.personalfinancetracker.dto.response.GoalExpenseResponseDto;
 import com.nirmal.personalfinancetracker.enums.ExpenseEnum;
 import com.nirmal.personalfinancetracker.enums.RecurrenceEnum;
 import com.nirmal.personalfinancetracker.model.Expense;
+import com.nirmal.personalfinancetracker.model.Goal;
 import com.nirmal.personalfinancetracker.model.User;
 import com.nirmal.personalfinancetracker.repository.ExpenseRepository;
+import com.nirmal.personalfinancetracker.repository.GoalRepository;
 import com.nirmal.personalfinancetracker.repository.UserRepository;
 import com.nirmal.personalfinancetracker.service.ExpenseService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,6 +22,7 @@ import java.time.LocalDateTime;
 import java.time.Month;
 import java.time.temporal.TemporalAdjusters;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Optional;
 
@@ -29,11 +34,13 @@ public class ExpenseServiceImpl implements ExpenseService {
     @Autowired
     private UserRepository userRepository;
     @Autowired
+    private GoalRepository goalRepository;
+    @Autowired
     private BudgetLimitServiceImpl budgetLimitServiceImpl;
     @Autowired
     private GoalServiceImpl goalServiceImpl;
     @Override
-    public Pair<Expense,List<Boolean>> addExpense(AddExpenseDto addExpenseDto) {
+    public ExpenseResponseDto addExpense(AddExpenseDto addExpenseDto) {
         Optional<User> user = userRepository.findById(addExpenseDto.getUserId());
         Expense expense = new Expense();
         expense.setUser(user.get());
@@ -42,13 +49,13 @@ public class ExpenseServiceImpl implements ExpenseService {
         expense.setDescription(addExpenseDto.getDescription());
 
         expenseRepository.save(expense);
-
-        List<Boolean> checkLimitInInterval = checkLimitInInterval(expense);
-        Pair<Expense,List<Boolean>> expenseBooleanListPair = Pair.of(expense, checkLimitInInterval);
-        return expenseBooleanListPair;
+        ExpenseResponseDto expenseResponseDto = new ExpenseResponseDto();
+        expenseResponseDto.setExpense(expense);
+        expenseResponseDto.setOverLimit(createHashMap(expense));
+        return expenseResponseDto;
     }
 
-    public Pair<Expense, List<Boolean>> addExpense(int goalId, AddExpenseDto addExpenseDto) {
+    public GoalExpenseResponseDto addExpense(int goalId, AddExpenseDto addExpenseDto) {
         Optional<User> user = userRepository.findById(addExpenseDto.getUserId());
         Expense expense = new Expense();
         expense.setUser(user.get());
@@ -57,11 +64,14 @@ public class ExpenseServiceImpl implements ExpenseService {
         expense.setAmount(addExpenseDto.getAmount());
 
         expenseRepository.save(expense);
+        String data = goalServiceImpl.addAmountToGoal(expense.getAmount(), goalId);
+        GoalExpenseResponseDto goalExpenseResponseDto = new GoalExpenseResponseDto();
 
-        goalServiceImpl.addAmountToGoal(expense.getAmount(), goalId);
-        List<Boolean> checkLimitInInterval = checkLimitInInterval(expense);
-        Pair<Expense,List<Boolean>> expenseBooleanListPair = Pair.of(expense, checkLimitInInterval);
-        return expenseBooleanListPair;
+        goalExpenseResponseDto.setExpense(expense);
+        goalExpenseResponseDto.setOverLimit(createHashMap(expense));
+        goalExpenseResponseDto.setGoalId(goalId);
+        goalExpenseResponseDto.setData(data);
+        return goalExpenseResponseDto;
 
     }
     @Override
@@ -79,28 +89,75 @@ public class ExpenseServiceImpl implements ExpenseService {
     }
 
     @Override
-    public Pair<Expense,List<Boolean>> updateExpense(int expenseId, AddExpenseDto addExpenseDto) {
-        Expense expense = expenseRepository.findById(expenseId).get();
+    public ExpenseResponseDto updateExpense(int expenseId, AddExpenseDto addExpenseDto) {
         Optional<User> user = userRepository.findById(addExpenseDto.getUserId());
+
+        Expense expense = expenseRepository.findById(expenseId).get();
         expense.setUser(user.get());
         expense.setCategory(addExpenseDto.getCategory());
         expense.setDescription(addExpenseDto.getDescription());
         expense.setAmount(addExpenseDto.getAmount());
         expenseRepository.save(expense);
 
-        List<Boolean> checkLimitInInterval = checkLimitInInterval(expense);
-        Pair<Expense,List<Boolean>> expenseBooleanListPair = Pair.of(expense, checkLimitInInterval);
-        return expenseBooleanListPair;
+        ExpenseResponseDto expenseResponseDto = new ExpenseResponseDto();
+        expenseResponseDto.setExpense(expense);
+        expenseResponseDto.setOverLimit(createHashMap(expense));
+        return expenseResponseDto;
+    }
+
+    @Override
+    public ExpenseResponseDto updateGoalExpense(int goalId, int expenseId, AddExpenseDto addExpenseDto) {
+        Optional<Expense> expenseOptional = expenseRepository.findById(expenseId);
+        Optional<User> userOptional = userRepository.findById(addExpenseDto.getUserId());
+        Optional<Goal> goalOptional = goalRepository.findById(goalId);
+        ExpenseResponseDto expenseResponseDto = new ExpenseResponseDto();
+        if(expenseOptional.isPresent() && goalOptional.isPresent() && userOptional.isPresent()){
+
+            BigDecimal oldAmount = expenseOptional.get().getAmount();
+            BigDecimal newAmount = addExpenseDto.getAmount();
+            goalOptional.get().setAmountSaved(goalOptional.get().getAmountSaved().subtract(oldAmount).add(newAmount));
+            goalRepository.save(goalOptional.get());
+
+            expenseOptional.get().setUser(userOptional.get());
+            expenseOptional.get().setCategory(addExpenseDto.getCategory());
+            expenseOptional.get().setDescription(addExpenseDto.getDescription());
+            expenseOptional.get().setAmount(addExpenseDto.getAmount());
+            expenseRepository.save(expenseOptional.get());
+
+            expenseResponseDto.setExpense(expenseOptional.get());
+            expenseResponseDto.setOverLimit(createHashMap(expenseOptional.get()));
+            return expenseResponseDto;
+        }
+        return null;
     }
 
     @Override
     public String deleteExpense(int expenseId) {
         Optional<Expense> expense = expenseRepository.findById(expenseId);
         if(expense.isPresent()){
+            if(expense.get().getCategory()!=ExpenseEnum.GOAL){
+                expenseRepository.deleteById(expenseId);
+                return "success";
+            }
+            else {
+                return "goal id not given";
+            }
+        }
+        return "expense with id: "+expenseId+ " not present.";
+    }
+
+    @Override
+    public String deleteGoalExpense(int goalId, int expenseId) {
+        Optional<Expense> expense = expenseRepository.findById(expenseId);
+        Optional<Goal> goal = goalRepository.findById(goalId);
+        if(expense.isPresent() && goal.isPresent()){
+            BigDecimal amountReduced = expense.get().getAmount();
+            goal.get().setAmountSaved(goal.get().getAmountSaved().subtract(amountReduced));
+            goalRepository.save(goal.get());
             expenseRepository.deleteById(expenseId);
             return "success";
         }
-        return "expense with id: "+expenseId+ " not present.";
+        return "expense with id: "+expenseId+" or goal with id: "+ goalId +" not present";
     }
 
     public BigDecimal totalExpenseInInterval(ExpenseEnum category, RecurrenceEnum interval){
@@ -139,9 +196,7 @@ public class ExpenseServiceImpl implements ExpenseService {
     public List<Boolean> checkLimitInInterval(Expense expense){
         List<Boolean> checkLimitInInterval = new ArrayList<>();
         for(RecurrenceEnum recurrenceEnum: RecurrenceEnum.values()){
-            // expenses of category in interval
             BigDecimal totalExpenses = totalExpenseInInterval(expense.getCategory(),recurrenceEnum);
-            // check if already exceeded
             BigDecimal limit = budgetLimitServiceImpl.getLimit(expense.getCategory(),recurrenceEnum);
             if(limit!=null) {
                 if(totalExpenses.compareTo(limit) > 0){
@@ -156,5 +211,15 @@ public class ExpenseServiceImpl implements ExpenseService {
             checkLimitInInterval.add(null);
         }
         return checkLimitInInterval;
+    }
+
+    public HashMap<RecurrenceEnum,Boolean> createHashMap(Expense expense){
+        List<Boolean> checkLimitInInterval = checkLimitInInterval(expense);
+        RecurrenceEnum[] interval = RecurrenceEnum.values();
+        HashMap<RecurrenceEnum,Boolean> overLimit = new HashMap<>();
+        for(int i=0; i<interval.length;i++){
+            overLimit.put(interval[i], checkLimitInInterval.get(i));
+        }
+        return overLimit;
     }
 }
